// utils/enum.ts
export enum FileIdentifier {
  FACE_PHOTO = "face-photo",
  PASSPORT = "passport",
  INVITATION_LETTER = "invitation-letter",
  LEGAL_ENTRY = "legal-entry",
  EMPLOYMENT_CONTRACT = "employment-contract",
  JOB_DOCUMENTS = "job-documents",
  DIPLOMA_CERTIFICATE = "diploma-certificate",
  ADDITIONAL_1 = "additional-1",
  ADDITIONAL_2 = "additional-2"
}